+++ clientmanager/src/main/java/com/clientmanager/service/TeamServiceImpl.java
@@ -1,5 +1,6 @@
 package com.clientmanager.service;
 
+import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
 
@@ -9,18 +10,21 @@ import org.springframework.transaction.annotation.Transactional;
 
 import com.clientmanager.dao.PermissionDAO;
 import com.clientmanager.dao.TeamDAO;
+import com.clientmanager.exception.BusinessExcpetion;
 import com.clientmanager.model.Permission;
 import com.clientmanager.model.Team;
+import com.clientmanager.model.User;
+import com.clientmanager.model.Permission.PermissionType;
 
 @Service
 public class TeamServiceImpl implements TeamService {
-	
+
 	@Autowired
 	TeamDAO teamDAO;
-	
+
 	@Autowired
 	PermissionDAO permissionDAO;
-	
+
 	@Transactional
 	@Override
 	public List<Team> getAllTeams() {
@@ -29,32 +33,78 @@ public class TeamServiceImpl implements TeamService {
 
 	@Transactional
 	@Override
-	public Team getTeamById(int id) {
-		return teamDAO.findById(id).orElse(null);
+	public Team getTeamById(int id) throws BusinessExcpetion {
+		if (!teamDAO.existsById(id)) {
+			throw new BusinessExcpetion("team not found");
+		}
+		Team team = teamDAO.findById(id).orElse(null);
+		if (team == null || !team.validate()) {
+			throw new BusinessExcpetion("invalid team");
+		}
+		return team;
 	}
 
 	@Transactional
 	@Override
-	public Team createTeam(Team group) {
-		return teamDAO.save(group);
-	}
-
-	@Override
-	public Team updateTeam(Team group) {
-		if(teamDAO.existsById(group.getId())) {
-			return teamDAO.save(group);
-		} else {
-			return null;
+	public Team createTeam(Team team) throws BusinessExcpetion {
+		System.out.println("create team: " + team);
+		if (team == null || !team.validate()) {
+			throw new BusinessExcpetion("invalid team");
 		}
+		if (teamDAO.existsById(team.getId())) {
+			throw new BusinessExcpetion("duplicate team found");
+		}
+		return teamDAO.save(team);
 	}
 
 	@Override
-	public void deleteTeam(Team team) {		
+	public Team updateTeam(Team team) throws BusinessExcpetion {
+		if (team == null || !team.validate()) {
+			throw new BusinessExcpetion("invalid team");
+		}
+		if (!teamDAO.existsById(team.getId())) {
+			throw new BusinessExcpetion("team not found");
+		}
+		Team team2 = teamDAO.findById(team.getId()).orElse(null);
+		team2.setActive(team.isActive());
+		team2.setDescription(team.getDescription());
+		team2.setGroupname(team.getGroupname());
+//		team2.setGrouppermissions(team.getGrouppermissions());
+//		Set<Permission> savepermissions = team.getGrouppermissions();
+//		team.setGrouppermissions(new HashSet<Permission>());
+		Team ret = teamDAO.save(team2);
+//		ret.setGrouppermissions(savepermissions);
+		return ret;
+	}
+
+	@Override
+	public void deleteTeam(Team team) throws BusinessExcpetion {
+		System.out.println("deleteteam - 1");
+		if (team == null || !team.validate()) {
+			throw new BusinessExcpetion("invalid team");
+		}
+		System.out.println("deleteteam - 2");
+		if (!teamDAO.existsById(team.getId())) {
+			throw new BusinessExcpetion("team not found");
+		}
+		System.out.println("deleteteam - 3");
 		teamDAO.delete(team);
 	}
 
 	@Override
-	public Team addPermission(Team team, Permission permission) {
+	public Team addPermission(Team team, Permission permission) throws BusinessExcpetion {
+		if (team == null || !team.validate()) {
+			throw new BusinessExcpetion("invalid team");
+		}
+		if (permission == null || !permission.validate() || permission.getPermissiontype() != PermissionType.GROUP) {
+			throw new BusinessExcpetion("invalid permission");
+		}
+		if (!teamDAO.existsById(team.getId())) {
+			throw new BusinessExcpetion("team not found");
+		}
+		if (!permissionDAO.existsById(permission.getId())) {
+			throw new BusinessExcpetion("permission not found");
+		}
 		Set<Permission> permissions = team.getGrouppermissions();
 		permissions.add(permission);
 		team.setGrouppermissions(permissions);
@@ -63,13 +113,33 @@ public class TeamServiceImpl implements TeamService {
 	}
 
 	@Override
-	public Team removePermission(Team team, Permission permission) {
+	public Team removePermission(Team team, Permission permission) throws BusinessExcpetion {
+		if (team == null || !team.validate()) {
+			throw new BusinessExcpetion("invalid team");
+		}
+		if (permission == null || !permission.validate() || permission.getPermissiontype() != PermissionType.GROUP) {
+			throw new BusinessExcpetion("invalid permission");
+		}
+		if (!teamDAO.existsById(team.getId())) {
+			throw new BusinessExcpetion("team not found");
+		}
+		if (!permissionDAO.existsById(permission.getId())) {
+			throw new BusinessExcpetion("permission not found");
+		}
+		boolean found = false;
+		for (Permission perm : team.getGrouppermissions()) {
+			if (perm.getId() == permission.getId()) {
+				found = true;
+			}
+		}
+		if (!found) {
+			throw new BusinessExcpetion("permission not a member of team");
+		}
 		Set<Permission> permissions = team.getGrouppermissions();
 		permissions.remove(permission);
 		team.setGrouppermissions(permissions);
 		teamDAO.save(team);
 		return team;
 	}
-
 
 }
